fix bug

+++ hutool-core/src/main/java/cn/hutool/core/util/ByteUtil.java
@@ -341,7 +341,7 @@ public class ByteUtil {
 	 * @return float值
 	 * @since 5.7.18
 	 */
-	public static double bytesToFloat(final byte[] bytes) {
+	public static float bytesToFloat(final byte[] bytes) {
 		return bytesToFloat(bytes, DEFAULT_ORDER);
 	}
